requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.7.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
pathlib2>=2.3.7 